+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `__repr__` to `SetAttr`, `NoHandler`, `StreamWriterHandler`, `Logger`, `IndentLogger`, and `ChangeThreshold`
+
 ## [0.4.0] - 2022-06-06
 
 ### Added